// srm_ui/src/app/main/Models/etendering/ViewModels/terms-condition-view-model.ts
export class TermsConditionViewModel {
    id: string;
    RFQId: string;
    inputType: string;
    termsConditionsAttachmentPath: string;
    isEditable: boolean;
    beforeQuoteId: string;
    submitQuoteId: string;
    isRFQ: boolean;
    isRFI: boolean;
    isRFAQ: boolean;
    isPO: boolean;
    consents: any[];
    beforeQuoteStatus: string;
    submitQuoteStatus: string;
    termsConditions: string;
    etMediaId: string;
    fileName: string;
    wordData: string;
    isChecked: boolean;
    isOnLoadMandatory: boolean;
}